move & rename tests for find_wp_config_path
